% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmd.R
\name{create_document}
\alias{create_document}
\title{Create a New RMarkdown File}
\usage{
create_document(title, .fun = NULL, ...)
}
\arguments{
\item{...}{arbitrary metadata added to yaml header}
}
\value{

}
\description{
Create a New RMarkdown File
}
