% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read Website Metadata}
\usage{
read_metadata(url, output = c("list", "data.frame"), force = FALSE)
}
\arguments{
\item{url}{charcter url of website, or path to html document}

\item{output}{character defining desired output type. Must be one of:
\itemize{
  \item(list)
  \item{data.frame}
}}

\item{force}{logical force metadata retrieval where site does not
authorise scraping. This option should only be used if you are
the owner of the site.}
}
\value{
list or data.frame
}
\description{
Read Website Metadata
}
\examples{
read_metadata('https://www.w3.org/')
}
